First upload test

+++ backend/src/main/java/com/devsuperior/DSCommerce/services/ProductService.java
@@ -67,6 +67,9 @@ public class ProductService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Resource not found!");
+        }
         try{
             repository.deleteById(id);
         }
